deloton_stuff/requirements.txt
pandas
numpy
confluent_kafka
psycopg2-binary
python-dotenv